+++ src/integrations/playwright/index.ts
@@ -1,0 +1,2 @@
+export { playwrightToKualitee } from './playwrightToKualitee';
+export { postPlaywrightReportOnKualitee } from './postPlaywrightReportOnKualitee';
